Renamed SpecialResources folder to OtherResources.
